+++ CHANGELOG.rst
@@ -7,6 +7,15 @@ The format is based on
 and this project adheres to
 `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_
 
+
+[Unreleased] - 2023-10-02
+-------------------------
+
+Fixed
+^^^^^
+- Fixed the issue that ``loop_shortcut`` failed to update the model "returns".
+
+
 [0.1.1] - 2023-06-23
 --------------------